fix: workaround for reading body in HttpAuthenticationMechanism
See merge request kvasir/kvasir-server!67

* fix: workaround for reading body in HttpAuthenticationMechanism

Quarkus framework has no way of accessing body in a custom HttpAuthenticationMechanism. A GraphQL query posted via the request body, can both be read or write operations, so we need to access the body to determine if access can be granted based on the provided token. An additional GraphQLQueryChecker interface was introduced, which can be implemented by the policy agent plugins, and is called in the GraphQL endpoints explicitly (where the query is available and can be passed along).